feat(article-detail): 親からdataを呼ぶ設計へ修正
