+++ src/app/article-detail/article-detail-routing.module.ts
@@ -7,12 +7,6 @@ const routes: Routes = [
   {
     path: '',
     component: ArticleDetailComponent,
-    children: [
-      {
-        path: ':articleId',
-        component: ArticleComponent,
-      },
-    ],
   },
 ];
