+++ src/app/article-detail/article-detail/article-detail.component.html
@@ -1,8 +1,11 @@
-<div class="container">
+<div class="container" *ngIf="article$ | async as article; else loading">
   <div class="wrap">
     <div>
-      <router-outlet></router-outlet>
+      <app-article [article]="article"></app-article>
     </div>
-    <app-side [article]="article"></app-side>
+    <app-side></app-side>
   </div>
 </div>
+<ng-template #loading>
+  <mat-spinner></mat-spinner>
+</ng-template>
