+++ src/app/article-detail/article-detail/article-detail.component.scss
@@ -9,3 +9,7 @@
   gap: 24px;
   margin: 0 auto;
 }
+
+mat-spinner {
+  margin: 64px auto;
+}
