+++ src/app/article-detail/article/article.component.html
@@ -1,5 +1,5 @@
 <section>
-  <div class="article" *ngIf="article$ | async as article; else loading">
+  <div class="article" *ngIf="article">
     <div class="hero">
       <figure class="hero__thumbnail">
         <img [src]="article.thumbnailURL" alt="" />
@@ -119,8 +119,4 @@
       </a>
     </div>
   </div>
-
-  <ng-template #loading>
-    <mat-spinner></mat-spinner>
-  </ng-template>
 </section>
